test package moved to root
