+++ test/InMemoryTaskManagerTEST.java
@@ -1,6 +1,10 @@
-package test.java;
+package test;
 
 import org.junit.jupiter.api.Test;
+
+import java.util.HashMap;
+import java.util.Map;
+
 import org.junit.jupiter.api.Assertions;
 
 import main.java.model.Task;
@@ -322,7 +326,9 @@ class InMemoryTaskManagerTEST {
 
 	@Test
 	void getTask_taskMapDontContainsTaskId_failure() {
-		Assertions.assertThrows(Exception.class, () -> tm.getTask(1));
+
+		Task task = new Task(-1, "null", "fromGetTask", TaskProgress.DONE);
+		Assertions.assertEquals(task, tm.getTask(1));
 	}
 
 	/*
@@ -337,7 +343,9 @@ class InMemoryTaskManagerTEST {
 
 	@Test
 	void getMainTask_maintaskMapDontContainsTaskId_failure() {
-		Assertions.assertThrows(Exception.class, () -> tm.getMainTask(1));
+
+		Task task = new MainTask(-1, "null", "fromGetMainTask");
+		Assertions.assertEquals(task, tm.getMainTask(1));
 	}
 
 	/*
@@ -353,8 +361,9 @@ class InMemoryTaskManagerTEST {
 	}
 
 	@Test
-	void getSubTask_subtaskDoNotMapContainsTaskId_failure() {
-		Assertions.assertThrows(Exception.class, () -> tm.getSubTask(1));
+	void getSubTask_subtasMapkDoNotContainsTaskId_failure() {
+		Task task = new SubTask(-1, "null", "fromGetSubTask", -1, TaskProgress.NEW);
+		Assertions.assertEquals(task, tm.getSubTask(1));
 	}
 
 	/*
@@ -554,6 +563,7 @@ class InMemoryTaskManagerTEST {
 		Assertions.assertTrue(tm.getMainTask(maintaskId).getTaskProgress() == TaskProgress.IN_PROGRESS);
 	}
 
+	@Test
 	void checkTaskProgress_changeMainTaskprogressFromNEWToDONE_succes() throws Exception {
 		MainTask maintask = new MainTask("MainTask", "Discription");
 		int maintaskId = tm.addMainTask(maintask);
@@ -561,9 +571,10 @@ class InMemoryTaskManagerTEST {
 		int subtaskId = tm.addSubTask(subtask);
 		SubTask newSubTask = new SubTask(subtaskId, "newSubTask", "newDiscription", maintaskId, TaskProgress.DONE);
 		tm.updateSubTask(newSubTask);
-		Assertions.assertTrue(tm.getMainTask(maintaskId).getTaskProgress() == TaskProgress.DONE);
+		Assertions.assertEquals(TaskProgress.DONE, tm.getMainTask(maintaskId).getTaskProgress());
 	}
 
+	@Test
 	void checkTaskProgress_changeMainTaskprogressFromDONEToNEW_succes() throws Exception {
 		MainTask maintask = new MainTask("MainTask", "Discription");
 		int maintaskId = tm.addMainTask(maintask);
@@ -574,5 +585,124 @@ class InMemoryTaskManagerTEST {
 		SubTask newSubTask2 = new SubTask(subtaskId, "newSubTask", "newDiscription", maintaskId, TaskProgress.NEW);
 		tm.updateSubTask(newSubTask2);
 		Assertions.assertTrue(tm.getMainTask(maintaskId).getTaskProgress() == TaskProgress.NEW);
+	}
+
+	@Test
+	void deleteTaskById_FromHistory_succes() {
+		Task task1 = new Task("Task-1", "Discription", TaskProgress.NEW);
+		tm.addTask(task1);
+
+		MainTask task2 = new MainTask("Task-2", "Discription");
+		tm.addMainTask(task2);
+
+		SubTask task3 = new SubTask("Task-3", "Discription", task2.getId(), TaskProgress.NEW);
+		tm.addSubTask(task3);
+
+		// дублирование задач в начале истории
+		tm.getTask(task1.getId());
+		tm.getTask(task1.getId());
+
+		// дублирование задач в середине истории
+		tm.getMainTask(task2.getId());
+		tm.getMainTask(task2.getId());
+
+		// дублирование задач в конце истории
+		tm.getSubTask(task3.getId());
+		tm.getSubTask(task3.getId());
+
+		Assertions.assertEquals(3, tm.getHistory().size());
+		Assertions.assertEquals(task1, tm.getHistory().get(0));
+		Assertions.assertEquals(task2, tm.getHistory().get(1));
+		Assertions.assertEquals(task3, tm.getHistory().get(2));
+
+		// удаление задач
+		tm.deleteTaskById(task1.getId());
+		tm.deleteSubTaskById(task3.getId());
+		tm.deleteMainTaskById(task2.getId());
+
+		// история пуста
+		Assertions.assertEquals(0, tm.getHistory().size());
+	}
+
+	@Test
+	void asd() throws Exception {
+		Task task1 = new Task("Task-1", "Discription", TaskProgress.NEW);
+		Task task2 = new Task("Task-2", "Discription", TaskProgress.NEW);
+		Task task3 = new Task("Task-3", "Discription", TaskProgress.NEW);
+		tm.addTask(task1);
+		tm.addTask(task2);
+		tm.addTask(task3);
+
+		MainTask maintask1 = new MainTask("Maintask-1", "Discription");
+		MainTask maintask2 = new MainTask("Maintask-2", "Discription");
+		MainTask maintask3 = new MainTask("Maintask-3", "Discription");
+		tm.addMainTask(maintask1);
+		tm.addMainTask(maintask2);
+		tm.addMainTask(maintask3);
+
+		SubTask subtask1 = new SubTask("Subtask-1", "Discription", maintask1.getId(), TaskProgress.NEW);
+		SubTask subtask2 = new SubTask("Subtask-2", "Discription", maintask2.getId(), TaskProgress.NEW);
+		SubTask subtask3 = new SubTask("Subtask-3", "Discription", maintask3.getId(), TaskProgress.NEW);
+		tm.addSubTask(subtask1);
+		tm.addSubTask(subtask2);
+		tm.addSubTask(subtask3);
+
+		tm.getTask(task1.getId());
+		tm.getTask(task2.getId());
+		tm.getTask(task3.getId());
+
+		tm.getMainTask(maintask1.getId());
+		tm.getMainTask(maintask2.getId());
+		tm.getMainTask(maintask3.getId());
+
+		tm.getSubTask(subtask1.getId());
+		tm.getSubTask(subtask2.getId());
+		tm.getSubTask(subtask3.getId());
+
+		Assertions.assertEquals(9, tm.getHistory().size());
+		Assertions.assertEquals(task1, tm.getHistory().get(0));
+		Assertions.assertEquals(task2, tm.getHistory().get(1));
+		Assertions.assertEquals(task3, tm.getHistory().get(2));
+
+		Assertions.assertEquals(maintask1, tm.getHistory().get(3));
+		Assertions.assertEquals(maintask2, tm.getHistory().get(4));
+		Assertions.assertEquals(maintask3, tm.getHistory().get(5));
+
+		Assertions.assertEquals(subtask1, tm.getHistory().get(6));
+		Assertions.assertEquals(subtask2, tm.getHistory().get(7));
+		Assertions.assertEquals(subtask3, tm.getHistory().get(8));
+
+		Map<Integer, Task> tasks = new HashMap<>();
+		tasks.put(task1.getId(), task1);
+		tasks.put(task2.getId(), task2);
+		tasks.put(task3.getId(), task3);
+
+		Map<Integer, Task> mainTasks = new HashMap<>();
+		mainTasks.put(maintask1.getId(), maintask1);
+		mainTasks.put(maintask2.getId(), maintask2);
+		mainTasks.put(maintask3.getId(), maintask3);
+
+		Map<Integer, Task> subTasks = new HashMap<>();
+		subTasks.put(subtask1.getId(), subtask1);
+		subTasks.put(subtask2.getId(), subtask2);
+		subTasks.put(subtask3.getId(), subtask3);
+
+		// удаление задач списком
+		tm.deleteAllTasks();
+
+		// проверка удаления всех задач
+		Assertions.assertEquals(6, tm.getHistory().size());
+		Assertions.assertEquals(maintask1, tm.getHistory().get(0));
+		Assertions.assertEquals(maintask2, tm.getHistory().get(1));
+		Assertions.assertEquals(maintask3, tm.getHistory().get(2));
+		Assertions.assertEquals(subtask1, tm.getHistory().get(3));
+		Assertions.assertEquals(subtask2, tm.getHistory().get(4));
+		Assertions.assertEquals(subtask3, tm.getHistory().get(5));
+
+		// удаление списком всех главных задач
+		tm.deleteAllMainTasks();
+
+		// проверка удаления всех главных задач и их подзадач
+		Assertions.assertEquals(0, tm.getHistory().size());
 	}
 }
