+++ test/ManagersTEST.java
@@ -1,4 +1,4 @@
-package test.java;
+package test;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Assertions;
